Created utilities module; other small changes based on PR comments
